Simplify propagator
