+++ qutip/solver/scattering.py
